BAU: QoL improvements env vars, script names, etc.
Use renamed docker stages.

+++ rundocker.sh
@@ -9,31 +9,36 @@ fi
 
 BROWSER="${2:-chromium}"
 
-SP=false
 case "${ENVIRONMENT}" in
-  build-sp)
-    SP=true
+  dev-api)
+    TEST_MODE=API
+    AWS_PROFILE=di-auth-development-admin
+    ENVIRONMENT=dev
+    ;;
+  dev-ui)
+    TEST_MODE=UI
+    AWS_PROFILE=di-authentication-development-admin
+    ENVIRONMENT=dev
+    ;;
+  build-api)
+    TEST_MODE=API
+    AWS_PROFILE=gds-di-development-admin
+    ENVIRONMENT=build
+    ;;
+  build-ui)
+    TEST_MODE=UI
     AWS_PROFILE=di-authentication-build-admin
     ENVIRONMENT=build
     ;;
-  staging-sp)
-    SP=true
-    AWS_PROFILE=di-authentication-staging-admin
+  staging-api)
+    TEST_MODE=API
+    AWS_PROFILE=di-auth-staging-admin
     ENVIRONMENT=staging
     ;;
-  build)
-    AWS_PROFILE=gds-di-development-admin
-    ;;
-  staging)
-    AWS_PROFILE=di-auth-staging-admin
-    ;;
-  dev)
-    AWS_PROFILE=di-auth-development-admin
-    ;;
-  dev-sp)
-    SP=true
-    AWS_PROFILE=di-authentication-development-admin
-    ENVIRONMENT=dev
+  staging-ui)
+    TEST_MODE=UI
+    AWS_PROFILE=di-authentication-staging-admin
+    ENVIRONMENT=staging
     ;;
   *)
     echo "Unconfigured environment: $ENVIRONMENT"
@@ -41,13 +46,16 @@ case "${ENVIRONMENT}" in
     ;;
 esac
 
+[ -f api-env-override ] || touch api-env-override
+[ -f ui-env-override ] || touch ui-env-override
+
 export AWS_PROFILE
 source ./scripts/check_aws_creds.sh
-results_dir=/tmp/results/$(date +%Y-%m-%d-%H-%M-%S)
+results_dir=./tmp/results/$(date +%Y-%m-%d-%H-%M-%S)
 mkdir -p "${results_dir}"
 
-if [ "${SP}" = "true" ]; then
-  docker build . -t acceptance:latest --target secure-pipelines \
+if [ "${TEST_MODE}" = "UI" ]; then
+  docker build . -t acceptance:latest --target auth-ui-local \
     --build-arg "SELENIUM_BASE=selenium/standalone-${BROWSER}:132.0"
   docker run -p 4442-4444:4442-4444 --privileged \
     -e CODEBUILD_BUILD_ID=1 -e AWS_REGION="${AWS_REGION:-eu-west-2}" \
@@ -55,16 +63,15 @@ if [ "${SP}" = "true" ]; then
     -v "${results_dir}":/test/results \
     --env-file <(aws configure export-credentials --format env-no-export) \
     -it --rm --entrypoint /bin/bash --shm-size="2g" acceptance:latest /run-tests.sh
-else
-  ./run-acceptance-tests.sh -e "${ENVIRONMENT}"
+fi
 
-  docker build . -t acceptance:latest --target base \
+if [ "${TEST_MODE}" = "API" ]; then
+  docker build . -t acceptance:latest --target auth-api-local \
     --build-arg "SELENIUM_BASE=selenium/standalone-${BROWSER}:132.0"
 
   docker run -p 4442-4444:4442-4444 --privileged \
     -e PARALLEL_BROWSERS=1 \
     -v "${results_dir}":/test/acceptance-tests/target/cucumber-report \
-    -v "$(pwd)/generated.env:/test/.env" \
     --env-file <(aws configure export-credentials --format env-no-export) \
     -it --rm --entrypoint /bin/bash --shm-size="2g" \
     acceptance:latest /test/run-acceptance-tests.sh -s "${ENVIRONMENT}"
